renamed module to better reflect purpose
